Add board class and update main to show game generated
