+++ src/main.py
@@ -6,6 +6,7 @@ from configuration.configuration import *
 from file_manager.file_manager import *
 from menu import Menu
 from game import *
+from board import *
 
 
 def main_menu():
@@ -59,42 +60,19 @@ def game_menu():
 
 
 def game():
-    """
-         def print_code(self, number):
-        #It will be removed and defined in new class related with UI
-        if number is None:
-            return '*'
-        return str(number + 1)
 
-    def print_board(self, board):
-        #It will be removed and defined in new class related with UI#
-        out = '-----------------------\n'
-        y = 8
-        for matrix in xrange(81):
-            out += self.print_code(board[matrix]) + ' '
-            if (matrix % 3) == 2:
-                    out += "| "
-            if matrix == y:
-                    out += '\n'
-                    if ((matrix >= 19) and (matrix <= 27)) or ((matrix >= 45) and (matrix <= 54)):
-                            out += '------+-------+--------\n'
-                    y += 9
-        out += '-----------------------\n'
-        return out
-    """
     os.system('cls')
     print ('\n      Sudoku Game')
-    print ('      """""" """"\n\n')
+    print ('      """""" """"\n')
     sudoku_game = Game("easy")
-    #board = sudoku_game.generate_game()
-    # print(temp)
-    #print sudoku_game.print_board(board)
-    #c = 0
-    #for x in board:
-    #    if x:
-    #        c += 1
-    #c = 81 - c
-    #print("Missing numbers: %i" % c)
+    board = Board(sudoku_game.generate_game())
+    print board.print_board()
+    c = 0
+    for x in board.board:
+        if x:
+            c += 1
+    c = 81 - c
+    print("Missing numbers: %i" % c)
     raw_input('\n\nPress any key: ')
     game_menu()
